requirements.txt
flask
torch
transformers
sentence-transformers
faiss-cpu
pandas